third_party/blink/web_tests/virtual/lna-workers-disabled/external/wpt/fetch/local-network-access/shared-worker-fetch.https.window-expected.txt

This is a testharness.js-based test.
PASS local to local: success.
FAIL private to local: failure. assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
FAIL private to local: success. assert_equals: fetch error expected (undefined) undefined but got (string) "TypeError"
PASS private to private: success.
FAIL public to local: failed preflight. assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
FAIL public to local: success. assert_equals: fetch error expected (undefined) undefined but got (string) "TypeError"
FAIL public to private: failed preflight. assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
FAIL public to private: success. assert_equals: fetch error expected (undefined) undefined but got (string) "TypeError"
PASS public to public: success.
FAIL treat-as-public to local: failed preflight. assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
PASS treat-as-public to local: success.
FAIL treat-as-public to private: failed preflight. assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
FAIL treat-as-public to private: success. assert_equals: fetch error expected (undefined) undefined but got (string) "TypeError"
PASS treat-as-public to public: success.
Harness: the test ran to completion.
